fix(group-roster-table): display table pagination correctly
